requirements.txt
attrs==20.2.0
click==7.1.2
colour==0.1.5
Django==2.2.13
django-haystack==3.0
importlib-metadata==2.0.0
jieba==0.42.1
Markdown==3.3.1
mysqlclient==2.0.1
Pillow==8.0.0
Pygments==2.7.2
pytz==2020.1
sqlparse==0.4.1
Whoosh==2.7.4
zipp==3.3.0
